feat: add local and session types
