+++ src/popup/app/services/app-events.ts
@@ -1,4 +1,10 @@
-import { FilterEvent } from './types';
+export enum FilterEvent {
+  UserChange = 'userChange',
+  Loaded = 'dataLoaded',
+  Saved = 'dataSaved',
+  Deleted = 'dataDeleted',
+  Selected = 'dataSelected',
+}
 
 export default class FilterEventTarget extends EventTarget {
   notify(event: FilterEvent) {
